Registering MesosSummaryStore in constructor instead of StoreConfig
